# [no-release-notes] Document and refactor `skip.List`