--- a/go/store/skip/list_test.go
+++ b/go/store/skip/list_test.go
@@ -86,6 +86,125 @@ func TestMemoryFootprint(t *testing.T) {
 	assert.Equal(t, 20, sz)
 }
 
+func BenchmarkList(b *testing.B) {
+	b.Run("benchmark Get", func(b *testing.B) {
+		b.Run("n=64", func(b *testing.B) {
+			vals := randomInts(64)
+			l := NewSkipList(bytes.Compare)
+			for i := range vals {
+				l.Put(vals[i], vals[i])
+			}
+			b.ResetTimer()
+
+			for i := 0; i < b.N; i++ {
+				_, ok := l.Get(vals[i%64])
+				if !ok {
+					b.Fail()
+				}
+			}
+			b.ReportAllocs()
+		})
+		b.Run("n=512", func(b *testing.B) {
+			vals := randomInts(512)
+			l := NewSkipList(bytes.Compare)
+			for i := range vals {
+				l.Put(vals[i], vals[i])
+			}
+			b.ResetTimer()
+
+			for i := 0; i < b.N; i++ {
+				_, ok := l.Get(vals[i%512])
+				if !ok {
+					b.Fail()
+				}
+			}
+			b.ReportAllocs()
+		})
+		b.Run("n=4096", func(b *testing.B) {
+			vals := randomInts(4096)
+			l := NewSkipList(bytes.Compare)
+			for i := range vals {
+				l.Put(vals[i], vals[i])
+			}
+			b.ResetTimer()
+
+			for i := 0; i < b.N; i++ {
+				_, ok := l.Get(vals[i%4096])
+				if !ok {
+					b.Fail()
+				}
+			}
+			b.ReportAllocs()
+		})
+		b.Run("n=32768", func(b *testing.B) {
+			vals := randomInts(32768)
+			l := NewSkipList(bytes.Compare)
+			for i := range vals {
+				l.Put(vals[i], vals[i])
+			}
+			b.ResetTimer()
+
+			for i := 0; i < b.N; i++ {
+				_, ok := l.Get(vals[i%32768])
+				if !ok {
+					b.Fail()
+				}
+			}
+			b.ReportAllocs()
+		})
+	})
+	b.Run("benchmark Put", func(b *testing.B) {
+		b.Run("n=64", func(b *testing.B) {
+			vals := randomInts(64)
+			l := NewSkipList(bytes.Compare)
+			for i := 0; i < b.N; i++ {
+				j := i % 64
+				if j == 0 {
+					l.Truncate()
+				}
+				l.Put(vals[j], vals[j])
+			}
+			b.ReportAllocs()
+		})
+		b.Run("n=512", func(b *testing.B) {
+			vals := randomInts(512)
+			l := NewSkipList(bytes.Compare)
+			for i := 0; i < b.N; i++ {
+				j := i % 512
+				if j == 0 {
+					l.Truncate()
+				}
+				l.Put(vals[j], vals[j])
+			}
+			b.ReportAllocs()
+		})
+		b.Run("n=4096", func(b *testing.B) {
+			vals := randomInts(4096)
+			l := NewSkipList(bytes.Compare)
+			for i := 0; i < b.N; i++ {
+				j := i % 4096
+				if j == 0 {
+					l.Truncate()
+				}
+				l.Put(vals[j], vals[j])
+			}
+			b.ReportAllocs()
+		})
+		b.Run("n=32768", func(b *testing.B) {
+			vals := randomInts(32768)
+			l := NewSkipList(bytes.Compare)
+			for i := 0; i < b.N; i++ {
+				j := i % 32768
+				if j == 0 {
+					l.Truncate()
+				}
+				l.Put(vals[j], vals[j])
+			}
+			b.ReportAllocs()
+		})
+	})
+}
+
 func testSkipList(t *testing.T, compare ValueCmp, vals ...[]byte) {
 	src.Shuffle(len(vals), func(i, j int) {
 		vals[i], vals[j] = vals[j], vals[i]