Improves `skip.List` performance for large lists by increasing max tower height from 5 to 10, and decreasing height-promotion probabilities.

before:
```
BenchmarkList
BenchmarkList/benchmark_Get
BenchmarkList/benchmark_Get/n=64
BenchmarkList/benchmark_Get/n=64-12    	 6292303	       217.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkList/benchmark_Get/n=512
BenchmarkList/benchmark_Get/n=512-12   	 4007974	       311.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkList/benchmark_Get/n=4096
BenchmarkList/benchmark_Get/n=4096-12  	 2511993	       513.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkList/benchmark_Get/n=32768
BenchmarkList/benchmark_Get/n=32768-12 	 1265131	       869.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkList/benchmark_Put
BenchmarkList/benchmark_Put/n=64
BenchmarkList/benchmark_Put/n=64-12    	 5640423	       213.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkList/benchmark_Put/n=512
BenchmarkList/benchmark_Put/n=512-12   	 4248255	       271.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkList/benchmark_Put/n=4096
BenchmarkList/benchmark_Put/n=4096-12  	 2934957	       447.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkList/benchmark_Put/n=32768
BenchmarkList/benchmark_Put/n=32768-12 	 1643672	       750.1 ns/op	       8 B/op	       0 allocs/op
PASS
```

after
```
BenchmarkList
BenchmarkList/benchmark_Get
BenchmarkList/benchmark_Get/n=64
BenchmarkList/benchmark_Get/n=64-12    	 5316765	       216.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkList/benchmark_Get/n=512
BenchmarkList/benchmark_Get/n=512-12   	 4198186	       288.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkList/benchmark_Get/n=4096
BenchmarkList/benchmark_Get/n=4096-12  	 2714745	       466.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkList/benchmark_Get/n=32768
BenchmarkList/benchmark_Get/n=32768-12 	 1846780	       594.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkList/benchmark_Put
BenchmarkList/benchmark_Put/n=64
BenchmarkList/benchmark_Put/n=64-12    	 4557384	       281.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkList/benchmark_Put/n=512
BenchmarkList/benchmark_Put/n=512-12   	 3194649	       376.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkList/benchmark_Put/n=4096
BenchmarkList/benchmark_Put/n=4096-12  	 2577404	       472.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkList/benchmark_Put/n=32768
BenchmarkList/benchmark_Put/n=32768-12 	 1731177	       652.3 ns/op	      10 B/op	       0 allocs/op
PASS

```